Add orange GroupCard color option
int-7575
